+++ Sources/CodexBarCore/Providers/Providers.swift
@@ -100,6 +100,7 @@ public enum UsageProvider: String, CaseIterable, Sendable, Codable {
     case hyper
     case gitkraken
     case devpass
+    case llmman
 }
 
 // swiftformat:enable sortDeclarations